requirements.txt
# SpecCompress-India: Dependencies
# Tested: Python 3.10+ | CUDA 11.8+ | Colab T4

# Core ML
torch>=2.1.0
torchvision>=0.16.0

# Data
numpy>=1.24.0
h5py>=3.9.0
scipy>=1.11.0

# ML Baselines
scikit-learn>=1.3.0

# Config
pyyaml>=6.0

# Optional: real data download
requests>=2.31.0

# Plotting (optional, for viz scripts)
matplotlib>=3.7.0

# Dev
tqdm>=4.65.0
